Make moximager work again

+++ rtools_gui/moxtester/moximager.py
@@ -24,14 +24,6 @@ class MoxImager:
         os.dup2(process_pipe[1], 1)
         os.dup2(process_pipe[1], 2)
         os.dup2(uart_sock, 3)
-        if args[0] == '--deploy':
-            os.execl(
-                '/bin/sh', '/bin/sh', '-c',
-                ' '.join(['/usr/bin/strace',
-                self.resources.mox_imager_exec,
-                '-F', '3',
-                *args, '2>/tmp/imager-err', '>/tmp/imager'])
-            )
         os.execl(
             self.resources.mox_imager_exec,
             self.resources.mox_imager_exec,
@@ -79,9 +71,11 @@ class MoxImager:
 
     def stop(self):
         """Terminate execution"""
-        os.kill(self.pid, signal.SIGKILL)
+        exit_code = None
+        if self.pid:
+            os.kill(self.pid, signal.SIGKILL)
+            exit_code = os.waitpid(self.pid, 0)[1]
         self.pid = None
-        exit_code = os.waitpid(self.pid, 0)[1]
         self.moxtester._d.default_baudrate()
         self.pexpect.close()
         #self.moxtester.default()
